trip-chat-assistant/pyproject.toml
[project]
name = "chat-assistant"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk[a2a]>=1.20.0",
    "litellm>=1.80.7",
    "mcp-weather-server>=0.6.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]

[build-system]
requires = ["uv_build>=0.9.14,<0.10.0"]
build-backend = "uv_build"

[tool.ruff]
target-version = "py312"
extend-exclude = ["libs/tripsphere"]

[tool.ruff.lint]
select = ["E", "F", "B", "I"]

[tool.mypy]
strict = true
python_version = "3.12"

[tool.pyright]
typeCheckingMode = "strict"
include = ["src", "tests"]
